pyproject.toml
[tool.poetry]
name = "polocas-napadu-api"
version = "2.2.3"
description = "API defending database of Poločas nápadu, theatre improvisation group"
authors = ["Pavel Žák <pavel@zak.global>"]
license = "MIT"
readme = "README.md"
packages = [
  { include = "accounting" },
  { include = "api" },
  { include = "bands" },
  { include = "blog" },
  { include = "events" },
  { include = "emailing" },
  { include = "fields" },
  { include = "games" },
  { include = "gsuite" },
  { include = "images" },
  { include = "inspirations" },
  { include = "locations" },
  { include = "photos" },
  { include = "profiles" },
  { include = "shows" },
  { include = "theatre_sports" },
  { include = "tickets" },
  { include = "voting" },
  { include = "workshops" },
]
include = [
  "manage.py",
  "**/*.mo",
]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1.3"
django-extensions = "^3.1.5"
django-filter = "^21.1"
graphene-django = "^3.0.0"
pillow = "^9.3.0"
sentry-sdk = "^1.11.0"
django-storages = "^1.12.3"
boto3 = "^1.26.6"
psycopg2-binary = "^2.9.5"
django-admin-sso = "^4.0.0"
django-cors-headers = "^3.11.0"
qrcode = {extras = ["pil"], version = "^7.3.1"}
django-colorfield = "^0.7.2"
django-graphql-jwt = "^0.3.4"
django-admin-autocomplete-filter = "^0.7.1"
django-nested-admin = "^4.0.0"
fiobank = "^3.0.0"
django-searchable-encrypted-fields = "^0.1.9"
django-apptemplates = "^1.5"
czech-sort = "^1.0.1"
sorl-thumbnail = "^12.9.0"
gunicorn = "^20.1.0"
pytest-django = "^4.5.2"

[tool.poetry.group.dev.dependencies]
typing-extensions = "^4.4.0"
black = "^22.10.0"
pytest-freezegun = "^0.4.2"
pytest-watch = "^4.2.0"
pytest-cov = "^4.0.0"
prospector = "^1.7.7"
model-bakery = "^1.5.0"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
